requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==41.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
psycopg2==2.9.11
pycparser==3.0
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
